fix(tests): merge existing API auth tests into common tests
